Refactor - nameValidation 분리
